+++ chip-testing/src/scripts/Start.py
@@ -1,3 +1,0 @@
-#!/bin/sh
-echo "@ChipTestRunner: Command: 'Start' $@"
-sleep 3